requirements.txt
requests
playwright
pytest
python-dateutil